# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/courses": {
      "post": {
        "description": "Creates a new course record.",
        "operationId": "CoursesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created course.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a new course",
        "tags": [
          "courses"
        ]
      },
      "get": {
        "description": "Returns a list of all courses.",
        "operationId": "CoursesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "A list of courses.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve all courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/bulk": {
      "post": {
        "description": "Creates multiple course records in a single operation.",
        "operationId": "CoursesController_createMany",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateManyCoursesInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The number of courses created.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": {
                      "type": "number",
                      "description": "Number of courses created",
                      "example": 5
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create multiple courses",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}": {
      "get": {
        "description": "Returns a single course by its ID.",
        "operationId": "CoursesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the course",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The course with the specified ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "404": {
            "description": "Course not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Retrieve a course by ID",
        "tags": [
          "courses"
        ]
      },
      "put": {
        "description": "Updates an existing course record.",
        "operationId": "CoursesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the course",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The updated course.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Course not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Update a course by ID",
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "description": "Deletes a course record by its ID.",
        "operationId": "CoursesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the course",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Course successfully deleted."
          },
          "404": {
            "description": "Course not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a course by ID",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/relationship": {
      "post": {
        "description": "Creates a new relationship between courses.",
        "operationId": "CoursesController_createRelationship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCourseRelationshipInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The newly created course relationship.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CourseRelationshipResult"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Create a course relationship",
        "tags": [
          "courses"
        ]
      },
      "delete": {
        "description": "Deletes an existing relationship between courses.",
        "operationId": "CoursesController_deleteRelationship",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteCourseRelationshipInput"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Course relationship successfully deleted."
          },
          "404": {
            "description": "Course relationship not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Delete a course relationship",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}/modules": {
      "post": {
        "description": "Sets or updates the modules associated with a course.",
        "operationId": "CoursesController_setModules",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the course",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The course with the updated modules.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Course"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request."
          },
          "404": {
            "description": "Course or Module not found."
          },
          "500": {
            "description": "Internal Server Error."
          }
        },
        "summary": "Set modules for a course",
        "tags": [
          "courses"
        ]
      }
    },
    "/courses/{id}/summary": {
      "get": {
        "description": "Returns a human-readable plain text summary for the specified course.",
        "operationId": "CoursesController_getSummary",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "The ID of the course",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plain text summary of the course",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid ID format"
          },
          "404": {
            "description": "Course not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "summary": "Get course summary",
        "tags": [
          "courses"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lumina API",
    "description": "API documentation for Lumina server",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "Translation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "en_text": {
            "type": "string"
          },
          "he_text": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "en_text",
          "he_text"
        ]
      },
      "Faculty": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "description": {
            "$ref": "#/components/schemas/Translation"
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "degrees"
        ]
      },
      "Degree": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "facultyId": {
            "type": "string",
            "nullable": true
          },
          "faculty": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Faculty"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "courses",
          "facultyId",
          "faculty"
        ]
      },
      "University": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "courses": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Course"
            }
          },
          "degrees": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Degree"
            }
          }
        },
        "required": [
          "id",
          "name",
          "courses",
          "degrees"
        ]
      },
      "RelationshipMetadataKey": {
        "type": "string",
        "enum": [
          "REASON",
          "TYPE",
          "DESCRIPTION"
        ]
      },
      "RelationshipMetadata": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "key": {
            "allOf": [
              {
                "$ref": "#/components/schemas/RelationshipMetadataKey"
              }
            ]
          },
          "value": {
            "type": "string"
          },
          "blockRelationshipId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "key",
          "value",
          "blockRelationshipId"
        ]
      },
      "BlockRelationship": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteId": {
            "type": "string"
          },
          "postrequisiteId": {
            "type": "string"
          },
          "prerequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "postrequisite": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "metadata": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RelationshipMetadata"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteId",
          "postrequisiteId",
          "prerequisite",
          "postrequisite",
          "metadata"
        ]
      },
      "Block": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "prerequisiteFor": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "postrequisiteOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BlockRelationship"
            }
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          }
        },
        "required": [
          "id",
          "prerequisiteFor",
          "postrequisiteOf",
          "modules"
        ]
      },
      "QuestionValidationStatus": {
        "type": "string",
        "enum": [
          "ai_generated",
          "in_manual_review",
          "approved",
          "rejected"
        ]
      },
      "SelectAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "isCorrect": {
            "type": "boolean"
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "isCorrect",
          "translationId",
          "text",
          "answerId"
        ]
      },
      "UnitAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "unit": {
            "type": "string"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "unit",
          "answerId"
        ]
      },
      "NumberAnswer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "value": {
            "type": "number"
          },
          "answerId": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "value",
          "answerId"
        ]
      },
      "Answer": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "SelectAnswer": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectAnswer"
            }
          },
          "UnitAnswer": {
            "$ref": "#/components/schemas/UnitAnswer"
          },
          "NumberAnswer": {
            "$ref": "#/components/schemas/NumberAnswer"
          }
        },
        "required": [
          "id",
          "questionId"
        ]
      },
      "QuestionType": {
        "type": "string",
        "enum": [
          "selection",
          "value",
          "void"
        ]
      },
      "QuestionPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "questionId": {
            "type": "string"
          },
          "partQuestionId": {
            "type": "string"
          },
          "order": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "questionId",
          "partQuestionId",
          "order"
        ]
      },
      "Question": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "validationStatus": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionValidationStatus"
              }
            ]
          },
          "translationId": {
            "type": "string"
          },
          "text": {
            "$ref": "#/components/schemas/Translation"
          },
          "Answer": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Answer"
            }
          },
          "type": {
            "allOf": [
              {
                "$ref": "#/components/schemas/QuestionType"
              }
            ]
          },
          "Modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Parts": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          },
          "PartOf": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QuestionPart"
            }
          }
        },
        "required": [
          "id",
          "validationStatus",
          "translationId",
          "text",
          "Answer",
          "type",
          "Modules",
          "Parts",
          "PartOf"
        ]
      },
      "Module": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "translationId": {
            "type": "string"
          },
          "subModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "parentModules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          },
          "questions": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Question"
            }
          }
        },
        "required": [
          "id",
          "name",
          "translationId",
          "subModules",
          "parentModules",
          "Block",
          "questions"
        ]
      },
      "Course": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "$ref": "#/components/schemas/Translation"
          },
          "universityId": {
            "type": "string"
          },
          "university": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/University"
              }
            ]
          },
          "modules": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Module"
            }
          },
          "publishedAt": {
            "format": "date-time",
            "type": "string",
            "nullable": true
          },
          "Block": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/Block"
              }
            ]
          }
        },
        "required": [
          "id",
          "name",
          "universityId",
          "university",
          "modules",
          "publishedAt",
          "Block"
        ]
      },
...
      "CreateCourseInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Course name"
          },
          "universityId": {
            "type": "string",
            "description": "Owning university id",
            "format": "uuid"
          }
        },
        "required": [
          "name",
          "universityId"
        ]
      },
      "CreateManyCoursesInput": {
        "type": "object",
        "properties": {
          "courses": {
            "description": "Array of course data to create",
            "example": [
              {
                "name": "Computer Science",
                "universityId": "uuid1"
              },
              {
                "name": "Mathematics",
                "universityId": "uuid2"
              }
            ],
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateCourseInput"
            }
          }
        },
        "required": [
          "courses"
        ]
      },
      "CreateCourseRelationshipInput": {
        "type": "object",
        "properties": {
          "prerequisiteCourseId": {
            "type": "string",
            "description": "The ID of the prerequisite course"
          },
          "postrequisiteCourseId": {
            "type": "string",
            "description": "The ID of the postrequisite course"
          },
          "metadata": {
            "type": "object",
            "description": "Optional metadata for the relationship (JSON string)"
          }
        },
        "required": [
          "prerequisiteCourseId",
          "postrequisiteCourseId"
        ]
      },
      "CourseRelationshipResult": {
        "type": "object",
        "properties": {}
      },
      "DeleteCourseRelationshipInput": {
        "type": "object",
        "properties": {
          "prerequisiteCourseId": {
            "type": "string",
            "description": "The ID of the prerequisite course"
          },
          "postrequisiteCourseId": {
            "type": "string",
            "description": "The ID of the postrequisite course"
          }
        },
        "required": [
          "prerequisiteCourseId",
          "postrequisiteCourseId"
        ]
      },
...
    }
  }
}